kernel: add release status to kernel build
Patch-mainline: Never, SUSE specific
References: FATE#325281

We build many kernels from the same sources but not all packages are
destined to be officially released kernel.  This adds a new config
option to mark a kernel as released.  It defaults to no.  If it
is an unreleased kernel it will show " (unreleased)" after the version
string in Oopses and also show "No, Unreleased kernel" whenever
supported status is printed.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>

Patch-name: patches.suse/kernel-add-release-status-to-kernel-build.patch