Update stable patches to v5.10.98
Below patch is reverted:
  patches.stable/v5.10.97-0007-psi-Fix-uaf-issue-when-psi-trigger-is-destroyed-whil.patch
by:
  patches.suse/Revert-v5.10.97-psi-Fix-uaf-issue-when-psi-trigger-is-destroy.patch
then its mainline version is ported:
  patches.suse/v5.17-rc2-psi-Fix-uaf-issue-when-psi-trigger-is-destroyed-whil.patch
because we have backported this mainline patch before:
  patches.mainline/v5.16-cgroup-Allocate-cgroup_file_ctx-for-kernfs_open_file.patch

Below patch has patch header updated:
  patches.mainline/v5.16-cgroup-Allocate-cgroup_file_ctx-for-kernfs_open_file.patch

Below patches are refreshed:
  patches.openEuler/10644-psi-support-psi-under-cgroup-v1.patch
  patches.openEuler/12715-kabi-net-reserve-space-for-net-base-subsystem-relat.patch

Below patch is dropped:
  patches.openEuler/0027-ilp32-skip-ARM-erratum-1418040-for-ilp32-application.patch

The patch itself makes sense but it's not quite useful on SEL because
it's for aarch64 ILP32 mode, which we don't support. And it's a
openEuler downstream patch that no longer applies cleanly. Porting it to
SEL-2.0 kernel might be appropriate but that needs good understanding of
Arm detail. So remove it for now and someone else may bring it back
after careful review. See bsn#91.

Reference: bsn#91
Signed-off-by: Kai Liu <kai.liu@suse.com>
Change-Id: Ia249153a75a6d2e47a35d701f797bc92f2b99a21