packaging: build debuginfo when debuginfo flag is on
openEuler is using the %_enable_debug_packages macro to control whether
the debug{info,source} subpackages should be built. Check for it in
addition to the %__debug_package macro used on openSUSE, so the kconfig
option CONFIG_DEBUG_INFO will not be missed when debuginfo flag is on.

This fixes bsn#39.

References: bsn#39
Signed-off-by: Kai Liu <kai.liu@suse.com>
Change-Id: Ibfcb039fec3b1e6dd64e46d4259ecbf32acc0e3d