packaging: add the sel+leap153 flavor
This is a new flavor that builds the SEL kernel with the kconfigs from
openSUSE Leap 15.3.

The kconfigs for x86_64 and arm64 are taken from the SUSE kernel-source
tree, commit faf8ee0e427. Then they are refreshed by using below command:

./run_oldconfig.sh --flavor sel+leap153 -y

So that all new options take their defaults.

Below options are explicitly set to turn on MPTCP and RNBD.

  CONFIG_INET_MPTCP_DIAG=m
  CONFIG_MPTCP_IPV6=y
  CONFIG_BLK_DEV_RNBD=y

This flavor is meant for internal testing and verification by the team,
binary package will not be published.

Signed-off-by: Kai Liu <kai.liu@suse.com>
Change-Id: I488cccb79b547bfc669e4a6c3873325cb8cfd697