SEL-2.0: introduce crypto patches to improve the performance of getrandom
Per the article [1], the performance of getrandom can be improved a lot
with below patches.

patches.suse/v5.17-rc1-lib-crypto-blake2s-include-as-built-in.patch
patches.suse/v5.17-rc1-random-use-BLAKE2s-instead-of-SHA1-in-extraction.patch

And we have to import some crypto patches for blake2s because of the
dependency.

Reference: bsn#133 - improve random performance

[1]. https://www.phoronix.com/scan.php?page=news_item&px=Linux-getrandom-8450p

Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>
Change-Id: I93d3ebeb6c9ecbfee91ce59f7951c6e398a24692