SEL-2.0: adjust cma arm64 kconfigs
This patch adjusts CMA related kconfigs on ARM64: set CONFIG_CMA_AREAS
to 19 and CONFIG_CMA_SIZE_MBYTES to 128.

CONFIG_CMA_AREAS should be set to 19 when CONFIG_NUMA is set:

 config CMA_AREAS
         int "Maximum count of the CMA areas"
         depends on CMA
         default 19 if NUMA
         default 7

When CONFIG_CMA_SIZE_MBYTES is set to 64M or 0, we got these errors:

 cma: cma_alloc: alloc failed, req-size: 32 pages, ret: -12

This is reported in bsn#169.

Closes: bsn#169
Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Change-Id: I43bea7fe7c73f39a729e1634424cef33d0ca990a