SEL-2.0: remove some scsi patches
Some SCSI core patches are introduced with upgrade hpsa and mpt3sas,
which caused sata device such as sata cdrom can't be loaded, and other
potential problems for other SCSI devices if relevant patches in below
threads are not ported.

https://lore.kernel.org/all/20211012233558.4066756-2-bvanassche@acm.org/
https://lore.kernel.org/all/20211007202923.2174984-1-bvanassche@acm.org/

Let's be conservative to avoid touch SCSI core code, and remove those
patches which are introduced because of SCSI core changes.

Reference: bsn#222

Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>
Change-Id: Ib276ba6515fb1c0b5a6ec22f08898ef41142154a