kconfig: turn off CONFIG_BPFILTER
Recently we turned it on in 4d3191cbbf (kconfig: default flavor clean
up). But it caused numerous error messages in the kernel log, as below:

[   19.240224] bpfilter: Loaded bpfilter_umh pid 773
[   19.240425] Started bpfilter
[   19.241366] bpfilter: write fail -13
[   19.275231] bpfilter: Loaded bpfilter_umh pid 775
[   19.275463] bpfilter: write fail -13

The author suggested to turn it off as it's not ready for production
yet, according to [1]. So we turn it off.

[1]: https://www.spinics.net/lists/bpf/msg52138.html

The CONFIG_USERMODE_DRIVER option is autmatically turned off after
refreshing the kconfig, as it defaults to =n and was only enabled as the
dependency of CONFIG_BPFILTER.

References: bsn#267
Signed-off-by: Kai Liu <kai.liu@suse.com>
Change-Id: I5b8508512bad4087bea4fec588b0197b9ff55677